Move functions and files around
