+++ src/panels/Scene/SceneTree/FeaturedSceneTree.tsx
@@ -1,6 +1,5 @@
 import { Tree } from '@mantine/core';
 
-import { SceneTreeGroupPrefixKey } from '@/util/groupsHelpers';
 import {
   useGetAimNode,
   useGetAnchorNode,
@@ -8,7 +7,7 @@ import {
 } from '@/util/propertyTreeHooks';
 
 import { SceneTreeNode } from './SceneTreeNode';
-import { treeDataForSceneGraphNode } from './treeUtils';
+import { SceneTreeGroupPrefixKey, treeDataForSceneGraphNode } from './treeUtils';
 import { SceneTreeNodeData } from './types';
 
 /**
