+++ src/panels/Scene/SceneTree/SceneTreeNode.tsx
@@ -5,13 +5,13 @@ import { useWindowEvent } from '@mantine/hooks';
 import { CollapsableHeader } from '@/components/Collapsable/CollapsableHeader/CollapsableHeader';
 import { useAppDispatch, useAppSelector } from '@/redux/hooks';
 import { setSceneTreeSelectedNode } from '@/redux/local/localSlice';
-import { isGroupNode } from '@/util/groupsHelpers';
 
 import { useOpenCurrentSceneNodeWindow } from '../hooks';
 import { SceneGraphNodeHeader } from '../SceneGraphNode/SceneGraphNodeHeader';
 
 import { CurrentNodeView } from './CurrentNodeView';
 import { SceneTreeNodeData } from './types';
+import { isGroupNode } from './treeUtils';
 
 interface Props {
   node: SceneTreeNodeData;
