+++ src/panels/Scene/hooks.ts
@@ -1,17 +1,6 @@
 import React, { useCallback } from 'react';
 import { BoxData, PanelData, TabData } from 'rc-dock';
 
-import {
-  useGetBoolPropertyValue,
-  useGetFloatPropertyValue,
-  useOpenSpaceApi
-} from '@/api/hooks';
-import { Uri } from '@/types/types';
-import {
-  checkVisiblity,
-  enabledPropertyUri,
-  fadePropertyUri
-} from '@/util/propertyTreeHelpers';
 import { useWindowLayoutProvider } from '@/windowmanagement/WindowLayout/hooks';
 
 export const DefaultSgnWindowId = 'defaultSceneGraphNodeWindow';
@@ -66,33 +55,5 @@ export function useOpenCurrentSceneNodeWindow() {
   return {
     openCurrentNodeWindow,
     closeCurrentNodeWindow
-  };
-}
-
-export function usePropertyOwnerVisibility(uri: Uri) {
-  const luaApi = useOpenSpaceApi();
-
-  const [enabledPropertyValue, setEnabledProperty] = useGetBoolPropertyValue(
-    enabledPropertyUri(uri)
-  );
-  const [fadePropertyValue] = useGetFloatPropertyValue(fadePropertyUri(uri));
-  const isFadeable = fadePropertyValue !== undefined;
-
-  const isVisible = checkVisiblity(enabledPropertyValue, fadePropertyValue);
-
-  function setVisiblity(shouldShow: boolean, isImmediate: boolean = false) {
-    const fadeTime = isImmediate ? 0 : undefined;
-    if (!isFadeable) {
-      setEnabledProperty(shouldShow);
-    } else if (shouldShow) {
-      luaApi?.fadeIn(uri, fadeTime);
-    } else {
-      luaApi?.fadeOut(uri, fadeTime);
-    }
-  }
-
-  return {
-    isVisible,
-    setVisiblity
   };
 }
